Use TCP Segmentation
This commit is proposed to be a fix of #1.
Some DPIs drops IP fragmented packets so we are going to higher level
and fragmenting TCPs. Fragmented TCPs are normally handled by DPIs but
if they are going in reversed order everything is ok. In further reviews
of this technique may be a delay for 2nd packet should be implemented.